New feature to do signal plots with same size but without ratio below.
Changed the StackPlot::draw() method, such that the last argument is NOT a bool,
but is now a std::string, and must be one of "RATIO","SIG" or "NM1" (default is "RATIO").
RATIO - plots with ratio plot below, for W/Z background plots and internal signal plots.
SIG - plots like RATIO but without the ratio sub plot. Used for signal plots for paper.
NM1 - plots without ratio, and is short and wide. Used for N-1 plots.

All programs that use this method have been updated.